requirements.txt
requests==2.31.0
pillow==10.2.0
pytesseract==0.3.10
